backend/requirements-test.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pdfplumber==0.10.3
python-multipart==0.0.6
pytest==7.4.3
python-docx==0.8.11
docxtpl==0.16.7
pytest-cov==4.1.0
pytest-asyncio==0.21.1
pytest-mock==3.12.0
httpx==0.25.2
faker==20.1.0